Move attribute and etag logic into Repository
Breaks up the current `update` method into new `update` and `patch` methods.
(Breaking change)

- ETag/Version logic must be managed by Repositories, except for the case of `ServiceProviderConfiguration` (which is likely stored in memory)
  - Implementations should set `scimResource.meta.version`, how this version is defined is up to the implementation
- The `UpdateRequest` class has been removed, the `getPatchOperations` method has been moved into a new class `PatchGenerator`
- Current examples have been updated and demo simple in memory patch application. e.g. Get/apply patch/persist.
  - Other implementations may want generate update queries directly from the list of PatchOperations (skipping intermediate get step)
